+++ akar-core/src/akar/try_out.clj
@@ -6,4 +6,4 @@
   (:require [akar.patterns :refer :all])
   (:require [akar.combinators :refer :all])
   (:require [akar.syntax :refer :all])
-  (:require [n01se.syntax :refer [parse-forms syndoc]]))
+  (:require [panini.core :as panini]))